feat: public profile page, public habits
