+++ app/components/ProfileHeader.vue
@@ -1,6 +1,15 @@
+<script setup lang="ts">
+import type { User } from '~~/server/utils/db';
+
+export interface Props {
+  user?: User
+}
+
+defineProps<Props>()
+</script>
 <template>
   <div class="relative flex p-5">
-    <UserProfileInfo />
-    <ProfileActions />
+    <UserProfileInfo :user="user"/>
+    <ProfileActions v-if="!user"/>
   </div>
 </template>
