+++ app/router.options.ts
@@ -1,0 +1,5 @@
+import type { RouterOptions } from 'nuxt/schema'
+
+export default <RouterOptions>{
+    sensitive: true
+}
