+++ server/api/user/index.patch.ts
@@ -2,13 +2,9 @@ import { eq, and } from 'drizzle-orm';
 import { useValidatedParams, useValidatedBody, z, zh } from 'h3-zod';
 
 export default eventHandler(async event => {
-  const { id } = await useValidatedParams(event, {
-    id: zh.intAsString,
-  });
-
   const { user: requestUser } = await requireUserSession(event);
 
-  if (requestUser.id !== id) {
+  if (requestUser.id) {
     return createError({ statusCode: 401 })
   }
 
@@ -22,7 +18,7 @@ export default eventHandler(async event => {
   const user = await useDB()
     .update(tables.user)
     .set(updatedFields)
-    .where(and(eq(tables.user.id, id), eq(tables.user.id, requestUser.id)))
+    .where(and(eq(tables.user.id, requestUser.id), eq(tables.user.id, requestUser.id)))
     .returning()
     .get();
